Kodi 21 changes (videoversion support)
added the needed xml to support videoversion and extras per the docs.
It's pretty ugly, I'm not a designer, sorry.
Also there is a bug with togglebuttons and alttexturfocus, in that it just doesn't work for some reason.
Also, something in Kodi 21 beta 2 has refresh for movies broken but it works fine under tv.